# docs/openapi.json
{
  "components": {
    "schemas": {
      "AdminLoginRequest": {
        "properties": {
          "token": {
            "title": "Token",
            "type": "string"
          }
        },
        "required": [
          "token"
        ],
        "title": "AdminLoginRequest",
        "type": "object"
      },
...
      "CreateApiKeyRequest": {
        "properties": {
          "label": {
            "title": "Label",
            "type": "string"
          },
          "monthly_quota": {
            "default": 1000,
            "title": "Monthly Quota",
            "type": "integer"
          },
          "monthly_usd_quota": {
            "default": 5.0,
            "title": "Monthly Usd Quota",
            "type": "number"
          },
          "rpm_limit": {
            "default": 30,
            "title": "Rpm Limit",
            "type": "integer"
          }
        },
        "required": [
          "label"
        ],
        "title": "CreateApiKeyRequest",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "**한국어 사기 신호 검출 reference implementation** — 음성·텍스트·이미지·PDF·URL 입력에서 위험 신호를 검출하고 각 신호의 학술/법적 근거를 transparent 하게 보고한다.\n\n⚠️ **Identity Boundary** (CLAUDE.md): ScamGuardian 은 *사기 판정을 내리지 않는다*. VirusTotal 이 70개 백신의 검출 결과를 보고만 하는 모델과 동일. 판정 logic 은 통합한 기업(통신사·은행·메신저 앱)이 자기 risk tolerance 에 따라 구현한다.\n\n- **Public 4 endpoint** (`Public` 태그) — 외부 통합 시작점, `DetectionReport` 응답\n- **Webhook** (`Webhook` 태그) — 카카오 오픈빌더\n- **Admin** (3 태그) — 라벨링 / platform / training\n- **v4 draft** — Live Call Guard 설계 preview (구현 X)\n\n통합 가이드: [`docs/INTEGRATION_GUIDE.md`](https://github.com/example/scamguardian-v2/blob/main/docs/INTEGRATION_GUIDE.md)",
    "title": "ScamGuardian Signal Detection API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/admin/abuse-blocks": {
      "get": {
        "description": "abuse_guard 가 자동 블록한 user_id + 만료 시각.",
        "operationId": "admin_abuse_blocks_api_admin_abuse_blocks_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Admin Abuse Blocks Api Admin Abuse Blocks Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "DB 미설정 또는 유효성 실패"
          },
          "401": {
            "description": "어드민 토큰 누락 또는 무효"
          },
          "500": {
            "description": "서버 내부 오류"
          }
        },
        "summary": "현재 차단된 사용자 목록",
        "tags": [
          "Admin — Platform"
        ]
      }
    },
    "/api/admin/abuse-blocks/{user_id}/unblock": {
      "post": {
        "description": "자동 블록을 수동으로 해제. False positive 보정용.",
        "operationId": "admin_abuse_unblock_api_admin_abuse_blocks__user_id__unblock_post",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Admin Abuse Unblock Api Admin Abuse Blocks  User Id  Unblock Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "DB 미설정 또는 유효성 실패"
          },
          "401": {
            "description": "어드민 토큰 누락 또는 무효"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "description": "서버 내부 오류"
          }
        },
        "summary": "사용자 차단 해제",
        "tags": [
          "Admin — Platform"
        ]
      }
    },
    "/api/admin/api-keys": {
      "get": {
        "description": "모든 키 메타데이터(plaintext 제외) + 사용량 카운터.",
        "operationId": "admin_list_api_keys_api_admin_api_keys_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Admin List Api Keys Api Admin Api Keys Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "DB 미설정 또는 유효성 실패"
          },
          "401": {
            "description": "어드민 토큰 누락 또는 무효"
          },
          "500": {
            "description": "서버 내부 오류"
          }
        },
        "summary": "API key 목록",
        "tags": [
          "Admin — Platform"
        ]
      },
      "post": {
        "description": "외부 클라이언트용 `sg_<urlsafe>` API key 발급. **plaintext 는 응답에 1회만** 노출되며 DB 에는 sha256 해시만 저장된다.\n\n**Body** (`CreateApiKeyRequest`):\n- `label` — 클라이언트 식별용 (예: `\"discord-bot\"`)\n- `monthly_quota` — 월별 호출 수 (기본 1000)\n- `rpm_limit` — 분당 호출 제한 (기본 30)\n- `monthly_usd_quota` — 월별 비용 한도 USD (기본 5.0)",
        "operationId": "admin_create_api_key_api_admin_api_keys_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateApiKeyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Admin Create Api Key Api Admin Api Keys Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "DB 미설정 또는 유효성 실패"
          },
          "401": {
            "description": "어드민 토큰 누락 또는 무효"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "description": "서버 내부 오류"
          }
        },
        "summary": "API key 발급",
        "tags": [
          "Admin — Platform"
        ]
      }
    },
    "/api/admin/api-keys/{key_id}/revoke": {
      "post": {
        "description": "키 status 를 `revoked` 로 전환. 이후 모든 요청 401/403.",
        "operationId": "admin_revoke_api_key_api_admin_api_keys__key_id__revoke_post",
        "parameters": [
          {
            "in": "path",
            "name": "key_id",
            "required": true,
            "schema": {
              "title": "Key Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Admin Revoke Api Key Api Admin Api Keys  Key Id  Revoke Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "DB 미설정 또는 유효성 실패"
          },
          "401": {
            "description": "어드민 토큰 누락 또는 무효"
          },
          "404": {
            "description": "key not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "description": "서버 내부 오류"
          }
        },
        "summary": "API key revoke",
        "tags": [
          "Admin — Platform"
        ]
      }
    },
    "/api/admin/cost": {
      "get": {
        "description": "`cost_events` 테이블 ledger — provider × api_key × 일자별 USD 합계. Claude / OpenAI Whisper / Serper / VirusTotal 분리.\n\n**Query**: `days` (기본 30).",
        "operationId": "admin_cost_api_admin_cost_get",
        "parameters": [
          {
            "in": "query",
            "name": "days",
            "required": false,
            "schema": {
              "default": 30,
              "title": "Days",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Admin Cost Api Admin Cost Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "DB 미설정 또는 유효성 실패"
          },
          "401": {
            "description": "어드민 토큰 누락 또는 무효"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "description": "서버 내부 오류"
          }
        },
        "summary": "비용 집계 (cost ledger)",
        "tags": [
          "Admin — Platform"
        ]
      }
    },
    "/api/admin/login": {
      "post": {
        "description": "단일 admin token 검증. `SCAMGUARDIAN_ADMIN_TOKEN` 환경변수와 비교(`hmac.compare_digest`). 성공 시 호출자(Next.js)가 같은 값을 httpOnly 쿠키로 저장.\n\n**Body**: `{token: str}`. **이 엔드포인트만 admin 인증 면제** (검증 자체가 목적).",
        "operationId": "admin_login_api_admin_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Admin Login Api Admin Login Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "토큰 불일치"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "description": "ADMIN_TOKEN 미설정 — 어드민 비활성"
          }
        },
        "summary": "어드민 로그인",
        "tags": [
          "Admin — Platform"
        ]
      }
    },
...
    "/api/admin/observability": {
      "get": {
        "description": "`request_log` 테이블 — 모든 요청의 status / latency / error 기록.\n\n**Query**: `hours` (기본 24), `recent_limit` (기본 100).",
        "operationId": "admin_observability_api_admin_observability_get",
        "parameters": [
          {
            "in": "query",
            "name": "hours",
            "required": false,
            "schema": {
              "default": 24,
              "title": "Hours",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "recent_limit",
            "required": false,
            "schema": {
              "default": 100,
              "title": "Recent Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Admin Observability Api Admin Observability Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "DB 미설정 또는 유효성 실패"
          },
          "401": {
            "description": "어드민 토큰 누락 또는 무효"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "description": "서버 내부 오류"
          }
        },
        "summary": "Request log 요약 + 최근",
        "tags": [
          "Admin — Platform"
        ]
      }
    },
...
  },
...
}